pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mp-web-audit"
version = "0.1.0"
description = "CLI-basiertes Web-Auditing-Framework fuer genehmigte Sicherheitspruefungen"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "sslyze>=6.0",
    "python-nmap>=0.7",
    "typer[all]>=0.12",
    "rich>=13.0",
    "jinja2>=3.1",
    "pydantic>=2.5",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "respx>=0.21",
    "ruff>=0.5",
]

[project.scripts]
webaudit = "webaudit.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["src/webaudit"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
